Fix tests, building
