+++ package.json
@@ -4,12 +4,11 @@
   "description": "Google OAuth2 authentication client",
   "main": "dist/google-oauth2.js",
   "scripts": {
-    "build": "webpack --mode=production",
-    "build:test": "webpack --mode=development --progress",
-    "dev": "webpack --mode=development --watch --progress",
+    "build": "npm run clean && tsc",
+    "clean": "rimraf dist",
     "prepublishOnly": "npm run build",
-    "test": "npm run build:test && npm run test:specs",
-    "test:specs": "nyc mocha -r ./test/index.js ./test/specs/**/*.spec.js"
+    "test": "npm run build && npm run test:specs",
+    "test:specs": "nyc mocha ./test/specs/**/*.spec.js"
   },
   "files": [
     "dist/**/*.js",
@@ -33,17 +32,15 @@
   },
   "homepage": "https://github.com/buttercup/google-oauth2-client#readme",
   "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "babel-loader": "^8.0.6",
+    "@types/node": "^17.0.8",
     "chai": "^4.2.0",
     "eventemitter3": "^4.0.0",
     "mocha": "^6.2.0",
     "nyc": "^14.1.1",
     "query-string": "^6.8.1",
+    "rimraf": "^3.0.2",
     "sinon": "^7.3.2",
-    "webpack": "^4.36.1",
-    "webpack-cli": "^3.3.6"
+    "typescript": "^4.5.4"
   },
   "dependencies": {
     "cowl": "^1.0.0"
