--- test/index.js
@@ -1,7 +1,0 @@
-const sinon = require("sinon");
-const { expect } = require("chai");
-
-Object.assign(global, {
-    expect,
-    sinon
-});
